doc: update members list
* sync with with ncu-team sync

\#31 should have added @thelarkinn to the members list. It was closed
for unrelated reasons but we never got Sean on the list.

* s/Modules Team Members/Members

* remove hyperlink from "Members" header

The original hyperlink lead to:

https://github.com/orgs/nodejs/teams/modules/members

Which is only accessible to people in the Node.js org afaict. The
member list is in the doc, the link is likely to just confuse folks.

* update ncu-team meta data

It now can inline Members and Observers. This is connected to the ACL
in github. The Observers header is currently commented out.

* fix link in governance to membership list