doc: update membership list with ncu-team sync
As per the discussion in the last meeting I have reached out to
some members privately about moving to observer status. As people
respond and opt-in to being moved to observers I am updating their
status in the GitHub ACL.

This PR will remain open until the next meeting so that we can avoid multiple
PRs.

If anyone would like to be include in the list of people being moved
from Active-Member to Observer please comment in this thread and
I'll take care of that.

I wanted to thank all those included in this PR for their hard work
and willingness to move to be Observers. I hope that you all continue
to participate in the repo and attend meetings when it is convenient.