add soldair to observers
hey folks. o/
I'm working on defining a common base for user land loaders. My personal goal is to make loaders to speed up node startup time and make a nice way for these to be implemented. 

I'm excited to contribute and it was great meeting some of you at the collab summit.